Move deprecated FunctionHierarchy to interface
