+++ firedrake/__init__.py
@@ -40,7 +40,6 @@ from firedrake.linear_solver import *
 from firedrake.matrix_free.preconditioners import *
 from firedrake.mesh import *
 from firedrake.mg.mesh import *
-from firedrake.mg.function import *
 from firedrake.mg.functionspace import *
 from firedrake.mg.interface import *
 from firedrake.norms import *
