+++ firedrake/mg/__init__.py
@@ -1,6 +1,5 @@
 from __future__ import absolute_import
 
 from .mesh import *              # noqa: F401
-from .function import *          # noqa: F401
 from .functionspace import *     # noqa: F401
 from .interface import *         # noqa: F401
